+++ lib/src/optimizer.dart
@@ -413,10 +413,7 @@ class Optimizer implements Visitor<Context, Node> {
 
   @override
   TemplateNode visitTemplateNode(TemplateNode node, Context context) {
-    return node.copyWith(
-      blocks: visitNodes<Block>(node.blocks, context),
-      body: visitNode<Node>(node.body, context),
-    );
+    return node.copyWith(body: visitNode<Node>(node.body, context));
   }
 
   @override